refactor(polish): extract prompt composition into polish/prompt_compose.rs — behavior unchanged, 446 tests green
Pure move of 8 functions (context_premise, compose_polish/translate/qa prompts,
assemble_polish_system_prompt, hotword-block builders) out of polish.rs into a child
module. References the parent 'prompts' module + PolishSystemPromptAssembly via
'use super::*;'; pub(crate) fns re-exported via 'pub(crate) use prompt_compose::*;'.
No logic/signature/public-API changes; polish.rs 3436→3157 lines.

cargo check (+--tests): 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
